+++ CHANGES.md
@@ -17,6 +17,7 @@ The release versions are PyPi releases.
   * Remove most of `fake_filesystem_shutil` implementation, relying on the patched `os` module instead ([#194](../../issues/194)).
 
 #### Fixes
+  * Recursive rename was not prevented and caused invalid state ([#216](../../issues/216)).
   * `os.rmdir` failed with symlink ([#226](../../issues/226)).
   * Renaming with a target that has a regular file as parent raised incorrect exception ([#224](../../issues/224)).
   * Renaming a directory to an existing directory was not possible (Posix only, [#217](../../issues/217)).